+++ src/array.mli
@@ -257,8 +257,6 @@ module Permissioned : sig
   val unzip : ('a * 'b, [> read]) t -> ('a, [< _ perms]) t * ('b, [< _ perms]) t
   val sorted_copy : ('a, [> read]) t -> compare:('a -> 'a -> int) -> ('a, [< _ perms]) t
   val last : ('a, [> read]) t -> 'a
-  val empty : unit -> ('a, [< _ perms]) t
-  [@@deprecated "[since 2017-05] Use [ [||] ]"]
   val equal : ('a -> 'a -> bool) -> ('a, [> read]) t -> ('a, [> read]) t -> bool
   val unsafe_truncate : (_, [> write]) t -> len:int -> unit